Auto Voice Channel (#113)
* Creates a Members Sync command (#109)

This allows for the Guild Owner or Bot Owner (D0cR3d) to initiate a sync and log to the database of all users in the server. This WILL take a long time for large servers. There is a wait of 250ms between each logging in order to not DOS the systems.

* Updated Deps

* Added template for Warn command. Updated deps.

**Warn:**

- The warn command can now use optional note argument `clan` to have it warn use with pre-defined clan warning text. The word `clan` must be the only note keyword. Usage: `.warn User clan`.

**General:**

- Updated deps

* Moved file locations (#112)

**General:**

- Moved files from /lib into /util and fixed their references for better organization.

* General Development

- Created VoiceChannelManager
- Updated Events to read voice channel updates
- Updated SweeperClient to initialize new VoiceChannelManager
- Updated Constants file with three new constants.